refactor: move sonarr token resource objects to 'TmdbTokenResources.cs'
